+++ CHANGELOG.md
@@ -84,6 +84,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Share-one and widget_share_contact : Ugly contact selection Widget
 - Authentication_page : User can't go back to the authentication page if he is connected
 - OpenedListPage : Screen don't overflow anymore on wishlist name changing
+- openedListPage : Ugly padding in wishlist screen
 
 ### Security